Just return path if file already exists on heroku

+++ app/models/youtube_repository.rb
@@ -32,6 +32,7 @@ class YoutubeRepository
     yt = ViddlRb.get_urls_names("https://www.youtube.com/watch?v=#{video.uid}").first
     filename = yt[:name].gsub(/[^0-9A-Za-z.\-]/, '_')
     path = "public/videos/#{filename}"
+    return path if File.exist? path
     File.open(path, 'wb') do |video_file|
       open(yt[:url]) { |v| video_file.write v.read }
     end
